Pass 'application' as a tag if API version 5.4+
The problem appeared after updating Zabbix from 3 to 5.4 version.
As turned out Zabbix dropped possibility to pass application in
parameters since 5.4 was released. Hence there left lots of python
scripts that have:
```
conn.item.get(..., application="appname", ...)
```
Instead it still be possible to specify application name in tags like
below:
```
conn.item.get(..., tags=[{"tag": "Application", "value": "appname"}],
...)
```
It means that scripts using py-zabbix must be updated when they
switches to 5.4 version from lower one (if it contained passing of
application).
Changes in this commit provided checking if Zabbix API version 5.4 or
greater and if so the specified application would be passed as a tag.
So there is no need to update plenty of scripts.